+++ src/config/mod.rs
@@ -63,6 +63,8 @@ pub struct GeneralConfig {
     pub auto_update_install: bool,
     #[serde(default = "default_true")]
     pub shell_integration: bool,
+    #[serde(default = "default_true")]
+    pub desktop_notifications: bool,
 }
 
 impl Default for GeneralConfig {
@@ -74,6 +76,7 @@ impl Default for GeneralConfig {
             auto_update_check: default_true(),
             auto_update_install: false,
             shell_integration: default_true(),
+            desktop_notifications: default_true(),
         }
     }
 }